CHANGE: Modified results from omdb to contain ...
Modified results from omdb to contain a boolean. The boolean is supposed
to tell if the movie is in the users list.

POC:
Looked at the returned json file to see if the boolean had been added.
Forced a search which didn't have any response to test if this still
works. Succes